GR-74426: Split AnalyzeJavaHomeAccess support singleton
