+++ substratevm/src/com.oracle.svm.hosted/src/com/oracle/svm/hosted/AnalyzeJavaHomeAccessFeature.java
@@ -24,23 +24,11 @@
  */
 package com.oracle.svm.hosted;
 
-import java.util.Collections;
-import java.util.Set;
-import java.util.concurrent.ConcurrentSkipListMap;
-
 import org.graalvm.nativeimage.ImageSingletons;
 
 import com.oracle.svm.shared.feature.AutomaticallyRegisteredFeature;
 import com.oracle.svm.core.feature.InternalFeature;
 import com.oracle.svm.core.imagelayer.ImageLayerBuildingSupport;
-import com.oracle.svm.shared.singletons.ImageSingletonLoader;
-import com.oracle.svm.shared.singletons.ImageSingletonWriter;
-import com.oracle.svm.shared.singletons.LayeredPersistFlags;
-import com.oracle.svm.shared.singletons.traits.BuiltinTraits.BuildtimeAccessOnly;
-import com.oracle.svm.shared.singletons.traits.LayeredCallbacksSingletonTrait;
-import com.oracle.svm.shared.singletons.traits.SingletonLayeredCallbacks;
-import com.oracle.svm.shared.singletons.traits.SingletonLayeredCallbacksSupplier;
-import com.oracle.svm.shared.singletons.traits.SingletonTraits;
 import com.oracle.svm.shared.util.LogUtils;
 
 /**
@@ -49,57 +37,20 @@ import com.oracle.svm.shared.util.LogUtils;
  * output.
  */
 @AutomaticallyRegisteredFeature
-@SingletonTraits(access = BuildtimeAccessOnly.class, layeredCallbacks = AnalyzeJavaHomeAccessFeature.LayeredCallbacks.class)
 public class AnalyzeJavaHomeAccessFeature implements InternalFeature {
-    private boolean javaHomeUsed = false;
-    private Set<String> javaHomeUsageLocations = Collections.newSetFromMap(new ConcurrentSkipListMap<>());
-
-    public static AnalyzeJavaHomeAccessFeature instance() {
-        return ImageSingletons.lookup(AnalyzeJavaHomeAccessFeature.class);
-    }
-
-    public void setJavaHomeUsed() {
-        javaHomeUsed = true;
-    }
-
-    public boolean getJavaHomeUsed() {
-        return javaHomeUsed;
-    }
-
-    public void addJavaHomeUsageLocation(String location) {
-        javaHomeUsageLocations.add(location);
+    @Override
+    public void afterRegistration(AfterRegistrationAccess access) {
+        ImageSingletons.add(AnalyzeJavaHomeAccessSupport.class, new AnalyzeJavaHomeAccessSupport());
     }
 
     @Override
     public void beforeCompilation(BeforeCompilationAccess access) {
-        for (String location : javaHomeUsageLocations) {
+        AnalyzeJavaHomeAccessSupport support = AnalyzeJavaHomeAccessSupport.singleton();
+        for (String location : support.getJavaHomeUsageLocations()) {
             LogUtils.warning("System.getProperty(\"java.home\") detected at " + location);
         }
         if (!ImageLayerBuildingSupport.buildingSharedLayer()) {
-            javaHomeUsageLocations = null;
-        }
-    }
-
-    static class LayeredCallbacks extends SingletonLayeredCallbacksSupplier {
-        private static final String JAVA_HOME_USED = "javaHomeUsed";
-        private static final String JAVA_HOME_USAGE_LOCATIONS = "javaHomeUsageLocations";
-
-        @Override
-        public LayeredCallbacksSingletonTrait getLayeredCallbacksTrait() {
-            return new LayeredCallbacksSingletonTrait(new SingletonLayeredCallbacks<AnalyzeJavaHomeAccessFeature>() {
-                @Override
-                public LayeredPersistFlags doPersist(ImageSingletonWriter writer, AnalyzeJavaHomeAccessFeature singleton) {
-                    writer.writeInt(JAVA_HOME_USED, singleton.javaHomeUsed ? 1 : 0);
-                    writer.writeStringList(JAVA_HOME_USAGE_LOCATIONS, singleton.javaHomeUsageLocations.stream().toList());
-                    return LayeredPersistFlags.CALLBACK_ON_REGISTRATION;
-                }
-
-                @Override
-                public void onSingletonRegistration(ImageSingletonLoader loader, AnalyzeJavaHomeAccessFeature singleton) {
-                    singleton.javaHomeUsed = loader.readInt(JAVA_HOME_USED) == 1;
-                    singleton.javaHomeUsageLocations.addAll(loader.readStringList(JAVA_HOME_USAGE_LOCATIONS));
-                }
-            });
+            support.clearJavaHomeUsageLocations();
         }
     }
 }
